Fold single-consumer Health Check helpers into their call sites (LAZ-551)
Several helpers were exported from shared modules but only ever used by one
file, so the indirection cost a hop without buying reuse.

issueFor and the tracker-binding logic move into the tracking context, which was
their only caller; that leaves the tracker module holding just the factory, so
it is renamed from useHealthCheckTracking to healthCheckTracker to match. The
hook it was named for was a one-line useMemo with a single call site, now inline
in HealthCheckPage. countActiveIssues moves to the api layer that runs the scan.
modEntryId and isModHidden move into ModRequirementsContent, which was the only
consumer of modRequirementEntries, so that module goes.

No behaviour change.
